* Renamed User to SubscribedUser * Created POST request to add subscribed user, with email validation * Cleanup of pom.xml
