--- src/main/java/com/vejo/em/casa/be/entity/User.java
@@ -1,36 +1,0 @@
-package com.vejo.em.casa.be.entity;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "user")
-public class User {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @Column
-    private String email;
-
-    public User() {}
-    public User(String email) {
-        this.email = email;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
