--- src/main/java/com/vejo/em/casa/be/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.vejo.em.casa.be.repository;
-
-import com.vejo.em.casa.be.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long>, JpaSpecificationExecutor<User> {
-}
