+++ src/main/java/com/vejo/em/casa/be/rest/request/SubscribeUserRequest.java
@@ -1,0 +1,20 @@
+package com.vejo.em.casa.be.rest.request;
+
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotNull;
+import java.io.Serializable;
+
+public class SubscribeUserRequest implements Serializable {
+
+    @NotNull
+    @Email
+    private String email;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+}
