+++ src/main/java/com/vejo/em/casa/be/service/CategoryService.java
@@ -21,7 +21,6 @@ public class CategoryService {
         return repository.findById(id).orElse(null);
     }
 
-
     public Category save(Category category) {
         return repository.save(category);
     }
